+++ FemDesign.Core/FemDesign.Core.csproj
@@ -71,8 +71,8 @@
     <Compile Include="GenericClasses\IStageElement.cs" />
     <Compile Include="GenericClasses\NamedEntityBase.cs" />
     <Compile Include="GenericClasses\TextFormatting.cs" />
-    <Compile Include="Loads\LoadCombinationTableGeneral.cs" />
-    <Compile Include="Loads\NationalAnnexEnum.cs" />
+    <Compile Include="Utils\LoadCombinationTable.cs" />
+    <Compile Include="Utils\NationalAnnexEnum.cs" />
     <Compile Include="Releases\RigidityDataType0.cs" />
     <Compile Include="Results\Convert.cs" />
     <Compile Include="Results\Node\PointSupportReactionMinMax.cs" />
@@ -201,9 +201,8 @@
     <Compile Include="Loads\LoadBase.cs" />
     <Compile Include="Loads\LoadCase.cs" />
     <Compile Include="Loads\LoadCombinationCaseBase.cs" />
-    <Compile Include="Loads\LoadCombinationTable.cs" />
     <Compile Include="Loads\LoadCombination.cs" />
-    <Compile Include="Loads\LoadCombinationType.cs" />
+    <Compile Include="Utils\LoadCombinationType.cs" />
     <Compile Include="Loads\LoadGroupBase.cs" />
     <Compile Include="Loads\LoadGroupRelationship.cs" />
     <Compile Include="Loads\LoadGroupTable.cs" />
@@ -375,6 +374,7 @@
   <ItemGroup>
     <None Include="packages.config" />
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSHARP.Targets" />
   <ProjectExtensions>
     <VisualStudio AllowExistingFolder="true" />
