+++ FemDesign.Grasshopper/FemDesign.Grasshopper.csproj
@@ -182,7 +182,6 @@
     <Compile Include="Supports\PointSupportConstruct.cs" />
     <Compile Include="Supports\StiffnessPoints.cs" />
     <Compile Include="Utils\LoadGroupCombineGeneral.cs" />
-    <Compile Include="Utils\LoadGroupCombine.cs" />
     <Compile Include="Loads\LoadGroupConstruct.cs" />
     <Compile Include="Loads\LoadCategoryDatabaseDefault.cs" />
     <Compile Include="Loads\LoadCategoryDatabaseListLoadCategoryNames.cs" />
